requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
redis==5.0.7
django-cors-headers==4.3.1
drf-spectacular==0.27.2
Pillow==10.4.0
pytest==8.3.3
pytest-django==4.8.0
